+++ docs/CHANGELOG.md
@@ -15,6 +15,7 @@ sebeolsik-sin1995 -> sebeolsik-3sin-1995
 ### Improve
 
 * Add `TreatJongseongAsChoseongCompose` addon [#332](https://github.com/Riey/kime/issues/332)
+* Fix 3-91 layout Z, S-Z bug [#335](https://github.com/Riey/kime/issues/335)
 
 ## 1.3.0