+++ docs/CHANGELOG.md
@@ -4,6 +4,12 @@
 
 ### Breaking
 
+### Improve
+
+## 1.3.1
+
+### Breaking
+
 * Rename sebeolsik layouts
 
 ```txt